+++ clang-tools-extra/clang-tidy/bsl/BslTidyModule.cpp
@@ -9,8 +9,9 @@
 #include "../ClangTidy.h"
 #include "../ClangTidyModule.h"
 #include "../ClangTidyModuleRegistry.h"
-#include "BooleanOperatorsForbiddenCheck.h"
 #include "AssignOpDeclRefQualifierCheck.h"
+#include "AutoTypeUsageCheck.h"
+#include "BooleanOperatorsForbiddenCheck.h"
 #include "ClassBaseCheck.h"
 #include "ClassFinalFunctionCheck.h"
 #include "ClassMemberInitCheck.h"
@@ -64,10 +65,12 @@ namespace bsl {
 class BslModule : public ClangTidyModule {
 public:
   void addCheckFactories(ClangTidyCheckFactories &CheckFactories) override {
-    CheckFactories.registerCheck<BooleanOperatorsForbiddenCheck>(
-        "bsl-boolean-operators-forbidden");
     CheckFactories.registerCheck<AssignOpDeclRefQualifierCheck>(
         "bsl-assign-op-decl-ref-qualifier");
+    CheckFactories.registerCheck<AutoTypeUsageCheck>(
+        "bsl-auto-type-usage");
+    CheckFactories.registerCheck<BooleanOperatorsForbiddenCheck>(
+        "bsl-boolean-operators-forbidden");
     CheckFactories.registerCheck<ClassBaseCheck>(
         "bsl-class-base");
     CheckFactories.registerCheck<ClassFinalFunctionCheck>(